Set ProjectID on Notification Delete Operations. (#329)
* Set ProjectID on Notification Delete Operations.

Previously the ProjectID env var would not be set on notification
delete operations, even when specified in the job args. Additionally,
the notification delete job args have been changed to use ProjectID
from PubSubStatus instead spec to be consistent with create.

* Sort EnvVars in notification job comparisons.

This allows the test to ignore differences in EnvVar order.

* Make ProjectID a required argument for delete in addition to create.

Since ProjectID is already required for create, it should be safe to
require it on delete as well.