Update webhook to latest style
Mostly copied from eventing's webhook.

Because the webhook config is now expressed as yaml, we can't insert the
system namespace into the webhook config name as easily as before. That
feature of the webhook has been removed. This means on upgrade there
will be an obsolete webhook configuration. We'll need to ask users to
delete that or delete it in code.