Add Delivery microbenchmarks with fake HTTP clients
Using a fake HTTP client allows greater isolation of the performance of the delivery processor and
more useful profiling data. The real client benchmarks are kept to detect performance differences in
the HTTP client itself as well as performance considerations that are only relevant with a real
client such as HTTP connection reuse.