feat: register pubsub client in broker and trigger reconciler if not passed (#1860)
* feat: register pubsub client in broker and trigger reconciler if not passed

Before this commit, when the pubsub client creation fails in the
controller while creating the reconciler, the reconciler will
create a new pubsub client in every reconciliation. Since the broker
and trigger only works with one project, we can reuse the client.
So after this commit, the newly created client will be registered.
This commit also refactors the reconcilers so that more common
code is moved to the getClientOrCreateNew function.

* add client failure test for broker

* add nil client creation test for trigger

* Fix comment